requirement.txt
Django 3.2.12
djangorestframework